much refactoring, add comments

+++ FormBE.Persistence/Model/FieldGroupField.cs
@@ -2,8 +2,8 @@
 
 public class FieldGroupField
 {
-    public int Id { get; set; }
-    public int FieldGroupId { get; set; }
+    public long Id { get; set; }
+    public long FieldGroupId { get; set; }
     public int FieldId { get; set; }
     public required FieldGroup FieldGroup { get; set; }
     public required Field Field { get; set; }
