+++ FormBE.Persistence/Model/FieldResponse.cs
@@ -2,8 +2,8 @@
 
 public class FieldResponse
 {
-    public int Id { get; set; }
-    public int FieldId { get; set; }
+    public long Id { get; set; }
+    public long FieldId { get; set; }
     public required string TelephoneNumber { get; set; }
     public required string Value { get; set; }
     public Instant SubmittedAt { get; set; }
