+++ FormBE.Persistence/Model/Form.cs
@@ -2,8 +2,8 @@
 
 public class Form
 {
-    public int Id { get; set; }
-    public int? GroupId { get; set; }
+    public long Id { get; set; }
+    public long? GroupId { get; set; }
     public required string? Name { get; set; }
     public Group? Group { get; set; }
     public required List<FormFieldGroup> FormFieldGroups { get; set; }
