+++ FormBE.Persistence/Repositories/FieldTypeRepository.cs
@@ -5,9 +5,32 @@ namespace FormBE.Persistence.Repositories;
 
 public interface IFieldTypeRepository
 {
-    public ValueTask<FieldType?> GetFieldTypeByIdAsync(int fieldTypeId, bool tracking = true, CancellationToken cancellationToken = default);
+    /// <summary>
+    /// Get a field type by its id.
+    /// </summary>
+    /// <param name="fieldTypeId">The id of the field tpye to get.</param>
+    /// <param name="tracking">If EF Core should track the entity.</param>
+    /// <param name="cancellationToken">A <see cref="CancellationToken"/> to observe while waiting for the task to complete.</param>
+    /// <returns>The specified field type or null if not found.</returns>
+    public ValueTask<FieldType?> GetFieldTypeByIdAsync(long fieldTypeId, bool tracking = true, CancellationToken cancellationToken = default);
+    
+    /// <summary>
+    /// Get all field types without tracking.
+    /// </summary>
+    /// <param name="cancellationToken">A <see cref="CancellationToken"/> to observe while waiting for the task to complete.</param>
+    /// <returns>All field types without tracking.</returns>
     public ValueTask<IReadOnlyCollection<FieldType>> GetAllFieldTypes(CancellationToken cancellationToken = default);
+    
+    /// <summary>
+    /// Begins tracking of a field type.
+    /// </summary>
+    /// <param name="fieldType">The field type to add.</param>
     public void AddFieldType(FieldType fieldType);
+    
+    /// <summary>
+    /// Begins to track a field type with the <see cref="EntityState.Deleted"/> state.
+    /// </summary>
+    /// <param name="fieldType">The field type to delete.</param>
     public void RemoveFieldType(FieldType fieldType);
 }
 
@@ -16,7 +39,7 @@ internal class FieldTypeRepository(DbSet<FieldType> fieldTypes) : IFieldTypeRepo
     private IQueryable<FieldType> FieldTypes => fieldTypes;
     private IQueryable<FieldType> NoTracking => FieldTypes.AsNoTracking();
     
-    public async ValueTask<FieldType?> GetFieldTypeByIdAsync(int fieldTypeId, bool tracking = true, CancellationToken cancellationToken = default)
+    public async ValueTask<FieldType?> GetFieldTypeByIdAsync(long fieldTypeId, bool tracking = true, CancellationToken cancellationToken = default)
     {
         IQueryable<FieldType> query = FieldTypes;
